Move read_to_class to helpers
