+++ plantcv/geospatial/read/geotif.py
@@ -7,7 +7,7 @@ import fiona
 from rasterio.mask import mask
 from plantcv.plantcv import warn, params, fatal_error
 from plantcv.plantcv._debug import _debug
-from plantcv.geospatial.read.netcdf import _read_to_class
+from plantcv.geospatial._helpers import _read_to_class
 from shapely.geometry import shape, MultiPoint, mapping
